Rename: replace IterableInterval with II
