+++ src/main/java/net/imagej/ops/copy/CopyII.java
@@ -28,51 +28,63 @@
  * #L%
  */
 
-package net.imagej.ops.map;
+package net.imagej.ops.copy;
 
 import net.imagej.ops.Contingent;
 import net.imagej.ops.Ops;
-import net.imglib2.Cursor;
+import net.imagej.ops.special.computer.Computers;
+import net.imagej.ops.special.computer.UnaryComputerOp;
+import net.imagej.ops.special.function.Functions;
+import net.imagej.ops.special.function.UnaryFunctionOp;
+import net.imagej.ops.special.hybrid.AbstractUnaryHybridCF;
 import net.imglib2.IterableInterval;
-import net.imglib2.RandomAccess;
-import net.imglib2.RandomAccessibleInterval;
 import net.imglib2.util.Intervals;
 
-import org.scijava.Priority;
 import org.scijava.plugin.Plugin;
 
 /**
- * {@link MapComputer} from {@link IterableInterval} inputs to
- * {@link RandomAccessibleInterval} outputs.
- *
- * @author Martin Horn (University of Konstanz)
- * @author Christian Dietz (University of Konstanz)
- * @author Tim-Oliver Buchholz (University of Konstanz)
- * @param <EI> element type of inputs
- * @param <EO> element type of outputs
+ * Copies an {@link IterableInterval} into another {@link IterableInterval}
+ * 
+ * @author Christian Dietz, University of Konstanz
+ * @param <T>
  */
-@Plugin(type = Ops.Map.class, priority = Priority.LOW_PRIORITY)
-public class MapIterableIntervalToRAI<EI, EO> extends
-	AbstractMapComputer<EI, EO, IterableInterval<EI>, RandomAccessibleInterval<EO>>
-	implements Contingent
-{
+@Plugin(type = Ops.Copy.IterableInterval.class, priority = 1.0)
+public class CopyII<T> extends
+		AbstractUnaryHybridCF<IterableInterval<T>, IterableInterval<T>> implements
+		Ops.Copy.IterableInterval, Contingent {
+
+	// used internally
+	private UnaryComputerOp<IterableInterval<T>, IterableInterval<T>> map;
+	private UnaryFunctionOp<IterableInterval<T>, IterableInterval<T>> imgCreator;
+
+	@SuppressWarnings({ "unchecked", "rawtypes" })
+	@Override
+	public void initialize() {
+		map = Computers.unary(ops(), Ops.Map.class, in(), in(), Computers.unary(
+			ops(), Ops.Copy.Type.class, in().firstElement().getClass(), in()
+				.firstElement().getClass()));
+		imgCreator = (UnaryFunctionOp) Functions.unary(ops(), Ops.Create.Img.class,
+			IterableInterval.class, in(), in().firstElement());
+	}
 
 	@Override
-	public void compute1(final IterableInterval<EI> input,
-		final RandomAccessibleInterval<EO> output)
-	{
-		final Cursor<EI> cursor = input.localizingCursor();
-		final RandomAccess<EO> rndAccess = output.randomAccess();
+	public IterableInterval<T> createOutput(final IterableInterval<T> input) {
+		// FIXME: Assumption here: Create an Img. I would rather like: Create
+		// what ever is best given the input.
+		return imgCreator.compute1(input);
+	}
 
-		while (cursor.hasNext()) {
-			cursor.fwd();
-			rndAccess.setPosition(cursor);
-			getOp().compute1(cursor.get(), rndAccess.get());
-		}
+	@Override
+	public void compute1(final IterableInterval<T> input,
+			final IterableInterval<T> output) {
+		map.compute1(input, output);
 	}
 
 	@Override
 	public boolean conforms() {
-		return out() == null || Intervals.equalDimensions(out(), in());
+		if (out() != null) {
+			return Intervals.equalDimensions(in(), out());
+		}
+		return true;
 	}
 }
