+++ src/main/java/net/imagej/ops/geom/CentroidII.java
@@ -27,69 +27,50 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
+package net.imagej.ops.geom;
 
-package net.imagej.ops.map;
-
-import net.imagej.ops.Contingent;
+import net.imagej.ops.Op;
 import net.imagej.ops.Ops;
-import net.imagej.ops.Parallel;
-import net.imagej.ops.special.computer.UnaryComputerOp;
-import net.imagej.ops.thread.chunker.ChunkerOp;
-import net.imagej.ops.thread.chunker.CursorBasedChunk;
+import net.imagej.ops.special.function.AbstractUnaryFunctionOp;
 import net.imglib2.Cursor;
 import net.imglib2.IterableInterval;
-import net.imglib2.RandomAccess;
-import net.imglib2.RandomAccessibleInterval;
-import net.imglib2.util.Intervals;
+import net.imglib2.RealLocalizable;
+import net.imglib2.RealPoint;
+import net.imglib2.roi.IterableRegion;
 
-import org.scijava.Priority;
 import org.scijava.plugin.Plugin;
 
 /**
- * Parallelized {@link MapComputer} from {@link IterableInterval} inputs to
- * {@link RandomAccessibleInterval} outputs.
+ * This {@link Op} computes the centroid of a {@link IterableRegion} (Label).
  * 
- * @author Christian Dietz (University of Konstanz)
- * @param <EI> element type of inputs
- * @param <EO> element type of outputs
+ * @author Tim-Oliver Buchholz, University of Konstanz.
  */
-@Plugin(type = Ops.Map.class, priority = Priority.LOW_PRIORITY + 2)
-public class MapIterableIntervalToRAIParallel<EI, EO> extends
-	AbstractMapComputer<EI, EO, IterableInterval<EI>, RandomAccessibleInterval<EO>>
-	implements Contingent, Parallel
-{
+@Plugin(type = Ops.Geometric.Centroid.class, priority = 1)
+public class CentroidII
+		extends
+			AbstractUnaryFunctionOp<IterableInterval<?>, RealLocalizable>
+		implements
+			Ops.Geometric.Centroid {
 
 	@Override
-	public void compute1(final IterableInterval<EI> input,
-		final RandomAccessibleInterval<EO> output)
-	{
-		ops().run(ChunkerOp.class, new CursorBasedChunk() {
-
-			@Override
-			public void execute(final int startIndex, final int stepSize,
-				final int numSteps)
-			{
-				final UnaryComputerOp<EI, EO> safe = getOp().getIndependentInstance();
-				final Cursor<EI> cursor = input.localizingCursor();
-
-				setToStart(cursor, startIndex);
-
-				final RandomAccess<EO> rndAccess = output.randomAccess();
-
-				int ctr = 0;
-				while (ctr < numSteps) {
-					rndAccess.setPosition(cursor);
-					safe.compute1(cursor.get(), rndAccess.get());
-					cursor.jumpFwd(stepSize);
-					ctr++;
-				}
+	public RealLocalizable compute1(final IterableInterval<?> input) {
+		int numDimensions = input.numDimensions();
+		double[] output = new double[numDimensions];
+		Cursor<?> c = input.localizingCursor();
+		while (c.hasNext()) {
+			c.fwd();
+			double[] pos = new double[numDimensions];
+			c.localize(pos);
+			for (int i = 0; i < output.length; i++) {
+				output[i] += pos[i];
 			}
-		}, input.size());
-	}
+		}
 
-	@Override
-	public boolean conforms() {
-		return out() == null || Intervals.equalDimensions(out(), in());
+		for (int i = 0; i < output.length; i++) {
+			output[i] = output[i] / input.size();
+		}
+
+		return new RealPoint(output);
 	}
 
 }
