+++ src/main/java/net/imagej/ops/image/ImageNamespace.java
@@ -221,12 +221,12 @@ public class ImageNamespace extends AbstractNamespace {
 	}
 
 	/** Executes the "invert" operation on the given arguments. */
-	@OpMethod(op = net.imagej.ops.image.invert.InvertIterableInterval.class)
+	@OpMethod(op = net.imagej.ops.image.invert.InvertII.class)
 	public <I extends RealType<I>, O extends RealType<O>> IterableInterval<O> invert(
 			final IterableInterval<O> out, final IterableInterval<I> in) {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<O> result = (IterableInterval<O>) ops().run(
-				net.imagej.ops.image.invert.InvertIterableInterval.class, out,
+				net.imagej.ops.image.invert.InvertII.class, out,
 				in);
 		return result;
 	}
@@ -239,7 +239,7 @@ public class ImageNamespace extends AbstractNamespace {
 		return ops().run(Ops.Image.Normalize.NAME, args);
 	}
 
-	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class)
+	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIIComputer.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> out, final IterableInterval<T> in)
@@ -247,12 +247,12 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops()
-				.run(net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class,
+				.run(net.imagej.ops.image.normalize.NormalizeIIComputer.class,
 					out, in);
 		return result;
 	}
 
-	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class)
+	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIIComputer.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> out, final IterableInterval<T> in,
@@ -261,12 +261,12 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class, out,
+				net.imagej.ops.image.normalize.NormalizeIIComputer.class, out,
 				in, sourceMin);
 		return result;
 	}
 
-	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class)
+	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIIComputer.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> out, final IterableInterval<T> in,
@@ -275,12 +275,12 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class, out,
+				net.imagej.ops.image.normalize.NormalizeIIComputer.class, out,
 				in, sourceMin, sourceMax);
 		return result;
 	}
 
-	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class)
+	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIIComputer.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> out, final IterableInterval<T> in,
@@ -289,12 +289,12 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class, out,
+				net.imagej.ops.image.normalize.NormalizeIIComputer.class, out,
 				in, sourceMin, sourceMax, targetMin);
 		return result;
 	}
 
-	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class)
+	@OpMethod(op = net.imagej.ops.image.normalize.NormalizeIIComputer.class)
 	public
 		<T extends RealType<T>>
 		IterableInterval<T>
@@ -304,13 +304,13 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalComputer.class, out,
+				net.imagej.ops.image.normalize.NormalizeIIComputer.class, out,
 				in, sourceMin, sourceMax, targetMin, targetMax);
 		return result;
 	}
 
 	@OpMethod(
-		op = net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class)
+		op = net.imagej.ops.image.normalize.NormalizeIIFunction.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> in)
@@ -318,13 +318,13 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class,
+				net.imagej.ops.image.normalize.NormalizeIIFunction.class,
 				in);
 		return result;
 	}
 
 	@OpMethod(
-		op = net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class)
+		op = net.imagej.ops.image.normalize.NormalizeIIFunction.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> in, final T sourceMin)
@@ -332,13 +332,13 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class,
+				net.imagej.ops.image.normalize.NormalizeIIFunction.class,
 				in, sourceMin);
 		return result;
 	}
 
 	@OpMethod(
-		op = net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class)
+		op = net.imagej.ops.image.normalize.NormalizeIIFunction.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> in, final T sourceMin, final T sourceMax)
@@ -346,13 +346,13 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class,
+				net.imagej.ops.image.normalize.NormalizeIIFunction.class,
 				in, sourceMin, sourceMax);
 		return result;
 	}
 
 	@OpMethod(
-		op = net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class)
+		op = net.imagej.ops.image.normalize.NormalizeIIFunction.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> in, final T sourceMin, final T sourceMax,
@@ -361,13 +361,13 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class,
+				net.imagej.ops.image.normalize.NormalizeIIFunction.class,
 				in, sourceMin, sourceMax, targetMin);
 		return result;
 	}
 
 	@OpMethod(
-		op = net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class)
+		op = net.imagej.ops.image.normalize.NormalizeIIFunction.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> in, final T sourceMin, final T sourceMax,
@@ -376,13 +376,13 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class,
+				net.imagej.ops.image.normalize.NormalizeIIFunction.class,
 				in, sourceMin, sourceMax, targetMin, targetMax);
 		return result;
 	}
 
 	@OpMethod(
-		op = net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class)
+		op = net.imagej.ops.image.normalize.NormalizeIIFunction.class)
 	public
 		<T extends RealType<T>> IterableInterval<T> normalize(
 			final IterableInterval<T> in, final T sourceMin, final T sourceMax,
@@ -391,7 +391,7 @@ public class ImageNamespace extends AbstractNamespace {
 		@SuppressWarnings("unchecked")
 		final IterableInterval<T> result =
 			(IterableInterval<T>) ops().run(
-				net.imagej.ops.image.normalize.NormalizeIterableIntervalFunction.class,
+				net.imagej.ops.image.normalize.NormalizeIIFunction.class,
 				in, sourceMin, sourceMax, targetMin, targetMax, isLazy);
 		return result;
 	}
@@ -407,7 +407,7 @@ public class ImageNamespace extends AbstractNamespace {
 	/** Executes the "project" operation on the given arguments. */
 	@OpMethod(ops = {
 			net.imagej.ops.image.project.DefaultProjectParallel.class,
-			net.imagej.ops.image.project.ProjectRAIToIterableInterval.class })
+			net.imagej.ops.image.project.ProjectRAIToII.class })
 	public <T, V> IterableInterval<V> project(final IterableInterval<V> out,
 			final RandomAccessibleInterval<T> in,
 			final UnaryComputerOp<Iterable<T>, V> method, final int dim) {
