+++ src/main/java/net/imagej/ops/map/MapRAIToII.java
@@ -32,68 +32,48 @@ package net.imagej.ops.map;
 
 import net.imagej.ops.Contingent;
 import net.imagej.ops.Ops;
-import net.imagej.ops.Parallel;
-import net.imagej.ops.special.computer.UnaryComputerOp;
-import net.imagej.ops.thread.chunker.ChunkerOp;
-import net.imagej.ops.thread.chunker.CursorBasedChunk;
 import net.imglib2.Cursor;
 import net.imglib2.IterableInterval;
+import net.imglib2.RandomAccess;
+import net.imglib2.RandomAccessibleInterval;
+import net.imglib2.util.Intervals;
 
 import org.scijava.Priority;
 import org.scijava.plugin.Plugin;
 
 /**
- * Parallelized {@link MapComputer} from {@link IterableInterval} inputs to
- * {@link IterableInterval} outputs. The {@link IterableInterval}s must have the
- * same iteration order.
- * 
+ * {@link MapComputer} from {@link RandomAccessibleInterval} input to
+ * {@link IterableInterval} output.
+ *
+ * @author Martin Horn (University of Konstanz)
  * @author Christian Dietz (University of Konstanz)
+ * @author Tim-Oliver Buchholz (University of Konstanz)
  * @param <EI> element type of inputs
  * @param <EO> element type of outputs
  */
-@Plugin(type = Ops.Map.class, priority = Priority.LOW_PRIORITY + 3)
-public class MapIterableIntervalToIterableIntervalParallel<EI, EO> extends
-	AbstractMapComputer<EI, EO, IterableInterval<EI>, IterableInterval<EO>>
-	implements Contingent, Parallel
+@Plugin(type = Ops.Map.class, priority = Priority.LOW_PRIORITY)
+public class MapRAIToII<EI, EO> extends
+	AbstractMapComputer<EI, EO, RandomAccessibleInterval<EI>, IterableInterval<EO>>
+	implements Contingent
 {
 
 	@Override
-	public boolean conforms() {
-		return out() == null || isValid(in(), out());
-	}
-
-	private boolean isValid(final IterableInterval<EI> input,
+	public void compute1(final RandomAccessibleInterval<EI> input,
 		final IterableInterval<EO> output)
 	{
-		return input.iterationOrder().equals(output.iterationOrder());
+		final Cursor<EO> cursor = output.localizingCursor();
+		final RandomAccess<EI> rndAccess = input.randomAccess();
+
+		while (cursor.hasNext()) {
+			cursor.fwd();
+			rndAccess.setPosition(cursor);
+			getOp().compute1(rndAccess.get(), cursor.get());
+		}
 	}
 
 	@Override
-	public void compute1(final IterableInterval<EI> input,
-		final IterableInterval<EO> output)
-	{
-		ops().run(ChunkerOp.class, new CursorBasedChunk() {
-
-			@Override
-			public void execute(final int startIndex, final int stepSize,
-				final int numSteps)
-			{
-				final UnaryComputerOp<EI, EO> safe = getOp().getIndependentInstance();
-
-				final Cursor<EI> inCursor = input.cursor();
-				final Cursor<EO> outCursor = output.cursor();
-
-				setToStart(inCursor, startIndex);
-				setToStart(outCursor, startIndex);
-
-				int ctr = 0;
-				while (ctr < numSteps) {
-					safe.compute1(inCursor.get(), outCursor.get());
-					inCursor.jumpFwd(stepSize);
-					outCursor.jumpFwd(stepSize);
-					ctr++;
-				}
-			}
-		}, input.size());
+	public boolean conforms() {
+		return out() == null || Intervals.equalDimensions(out(), in());
 	}
+
 }
