+++ src/test/java/net/imagej/ops/benchmark/AddOpBenchmarkTest.java
@@ -33,9 +33,9 @@ package net.imagej.ops.benchmark;
 import com.carrotsearch.junitbenchmarks.BenchmarkOptions;
 import com.carrotsearch.junitbenchmarks.BenchmarkRule;
 
-import net.imagej.ops.map.MapIterableIntervalInplaceParallel;
-import net.imagej.ops.map.MapIterableIntervalToIterableIntervalParallel;
-import net.imagej.ops.map.MapIterableIntervalToRAIParallel;
+import net.imagej.ops.map.MapIIInplaceParallel;
+import net.imagej.ops.map.MapIIToIIParallel;
+import net.imagej.ops.map.MapIIToRAIParallel;
 import net.imagej.ops.math.ConstantToArrayImage;
 import net.imagej.ops.math.ConstantToArrayImageP;
 import net.imagej.ops.math.ConstantToIIOutputII;
@@ -79,14 +79,14 @@ public class AddOpBenchmarkTest extends AbstractOpBenchmark {
 
 	@Test
 	public void fTestIterableIntervalMapperP() {
-		ops.run(MapIterableIntervalToIterableIntervalParallel.class, out, in, ops
+		ops.run(MapIIToIIParallel.class, out, in, ops
 			.op(NumericTypeBinaryMath.Add.class, null, NumericType.class,
 				new ByteType((byte) 10)));
 	}
 
 	@Test
 	public void fTestDefaultMapperP() {
-		ops.run(MapIterableIntervalToRAIParallel.class, out, in, ops.op(
+		ops.run(MapIIToRAIParallel.class, out, in, ops.op(
 			NumericTypeBinaryMath.Add.class, null, NumericType.class, new ByteType(
 				(byte) 10)));
 	}
@@ -99,7 +99,7 @@ public class AddOpBenchmarkTest extends AbstractOpBenchmark {
 
 	@Test
 	public void inTestDefaultInplaceMapperP() {
-		ops.run(MapIterableIntervalInplaceParallel.class, in, ops.op(
+		ops.run(MapIIInplaceParallel.class, in, ops.op(
 			AddConstantInplace.class, NumericType.class, new ByteType((byte) 10)));
 	}
